Keep PC in rdx for 8.6% on EXM
This delays writeback of PC into the Emu structure until guest exit, and
exploits my knowledge of the calling convention to arrange to keep it in
rdx throughout the interpreter loop, at least on AMD64 *ix. Which is
kind of fun.

Had to rewrite several of the op definitions to support this, and to
make that easier I've pulled support for the other interpreter models,
since they're now _real slow_ by comparison.